Refactor: new PackagingTestBase class.
